Fix interpretation of cycleLengths/availabilityFactors and add additional option to detailed input format
The interpretation of the cycle length in relation to the availability factor
was being used inconsistently before. This change makes them consistent and in
line with the rest of ARMI. Namely, the full cycle length is inclusive of the time
represented by the availability factor. Therefore the sum of each burn step length
will not equal the cycle length if availability factor is less than 1.

In addition, an option for specifying the cycle history in the detailed
cycles format has been added. Namely, the user can now also enter ``cycle length`` +
``burn steps`` within a given cycle, and that cycle will be divided into even increments.

These changes have been propagated through the tests and the user documentation.