Directory changers skip retrieval of missing files with a user warning. (#636)
Add check if the file that is being retrieved on a directory changer
exists and skip the copy if it does not exist. This is useful in instances
where optional files are requested when executing a physics code and
not all files are created because the user settings differ than the
default behavior.