Update control rod params & generalize the neutronics mesh converter to copy all multi-group results. (#613)
Adding a few control rod parameters and renaming others to support more general control rod plugin(s) for the framework. This also removes the getBlockLengthAboveAndBelowHeight because it was used by
the internal reactivity control plugin and now provides minimal value on the Assembly class.

Finally, the uniform mesh converter was improved on to copy all neutronics parameters back to the original core state compared to only copying back flux and mgFlux. This now uses the parameter system's multiGroupQuantities and detailedAxialExpansion categories on the block-level and the neutronics category on the reactor core level.