Initial commit for patcher utility
Wrote the skeleton for the patcher utility. The idea is to load in a
custom defined python module that has specifically named functions in
it. These functions are then run at certain times during the execution
of an ARMI case. These functions allow arbitrary python code to be
executed during the run, providing flexibility to the engineer to test
or update the run from a cetralized file.