Split `getCycleNode` into two functions with more clear purposes
`getCycleNode` had previously been getting the (cycle, node) tuple for a given
time node, but the docstring described it as getting the (cycle, node) for a given
time _step_. This was confusing. So the functionality for both options has been
moved to two separate functions, `getCycleNodeForCumulativeStep` and
`getCycleNodeForCumulativeNode`. That way the purpose of each is made very clear.

This change is then propagated out to the rest of the framework. In the process,
the failing historyTracker test has been fixed. Part of this involved removing
an assert statement that was checking `getCycleNode` in a very roundabout way
as an after-thought to the overall test. To make the test more explicit, tests
on the new `getCycleNodeForXXX` functions have been added directly.