Change `getTimeStepNum` to `getCumulativeNodeNum` and test it
The method `getTimeStepNum` had confusing semantics, because it referred to time
steps but really was getting time _nodes_. The name was thus changed to something
more indicative of its purpose.

In addition, the function was extended to handle the detailed cycles inputs.

Finally, a test has been added for the function.

This function was apparently unused in the framework, so no other pieces of code
had to be updated to align with this change.