Axial expansion update (#653)
To simplify the initialization of AxialExpansionChanger, removing the converterSettings object. This was used only to handle the detailedAxialExpansion setting. However, its use wasn't super clear. Instead, a detailed axial expansion instance variable (boolean) is used as a parameter. By default it's false.

Unit tests are updated to reflect this change as well as a previous update to a few class method names.