Add scope injection. Add tests.
While trying to test the module, I found that the scope needed to be
passed down to the patcher file in order to effect any lasting change to
objects in the primary scope. Since this ended up being a bit weirder
than the initial thoughts for this module, I put together the testing
suite with the intention that it would be a mini tutorial for how to
actually monkeypatch an ARMI job.