Try an implementation where the defaults are kept and the settings validator isn't used
For other reasons, I need to keep the defaults on the case settings of the simple
cycles settings. This means that I can't use the settings validator to screen
any more. Instead, I will try using the __getitem__ method of the `Settings` class
to test if a certain case setting can be gotten or not.

This will make it so that the simple cycles settings cannot be gotten (an error
will be raised) if the detailed cycles setting also exists. So in that case, not
only will plugins not be able to get those values, but also an operator will not
be able to be established.