Let's not delete a file inside a loop (v2) (#666)
The old pathTools() would repeatedly delete the same file/directory inside a loop. That seems suspect.

It also did not handle the situation well for when we want to delete a file/directory in an MPI or non-MPI scenario well.

This change to the method linearizes our file deletion, to only one delete call. It also gives the user control of whether or not this deletion should be MPI-safe or not.