Adding setting `inputHeightsConsideredHot` (#657)
Enable thermal expansion of assemblies when cores are built from blueprints + settings within armi.reactor.reactors.py::loadFromCs. The thermal expansion of each assembly is determined by the DeltaT of each component (DeltaT = Thot - Tinput).

To support backward compatibility (where no thermal expansion occurs) the setting, inputHeightsConsideredHot, was added. The default value for this is True. That way, current ARMI inputs retain their functionality. By setting to False, the armi.reactor.converters.axialExpansionChanger is spun up to expand each assembly as it is built.

To add a unit test that leverages this new capability (and that has the required dummy block for detailedAxialExpansion and the axialExpansionChanger), a new test reactor was created. The details:

    is based off of armi.tests.armiRun.yaml but sets:

axialExpansion: true
detailedAxialExpansion: true

    defines the required dummy block (dummy : &block_dummy) for armi.reactor.converters.axialExpansionChanger.py
    adds said dummy block to top of each assembly
        dummy block height = 10% of original assembly height (i.e., total assembly height w/out dummy block)