Fix setting the target component in axialExpansion (#677)
If more than one flag is in a block definition, the specifyTargetComponent method would fail with a Runtime error. The proposed fix is to loop the through a list of flags in order of importance break the loop once one of those flags matches. Still, this list of flags is incomplete so it is possible that componentWFlag would remain empty. In this case the original behavior is retained.

The unit test is a little redundant (testing that a RuntimeError is not raised, which is implicit in testing?) but I felt it important to be explicit there. Definitely willing to do something different.