Update documentation and modify check to be VARIANT-specific
Notes:

  * Upon inspecting the geodstCoordMap between VARIANT and Nodal, it was
noted that the VARIANT version is not ordered according to the four-color
pattern. Therefore, mention of four-color has been removed. This result is
a bit surprising, so a question regarding the cause of this difference
has been submitted to Mike Smith of ANL.

* The "iwnhfl" check should be VARIANT specific, so an additional boolean
check has been added. It has also been moved earlier in the __init__ to
throw an error earlier.