Fixing bug in memory profiler (#704)
Upon trying to use memory_profiler.py to debug a problem a user reported it was broken. Indeed, it appeared to have a couple of bugs in it. Furthermore, at least in Python 3.x, the third-party tool pympler.asizeof() was so slow as to be entirely unusualble.

To this end, I have fixed a couple of bugs in memory_profiler.py and removed all references to pympler.

My guess is that the Python standard library sys.getsizeof() method will under-report memory usage a bit. BUT it does work, and it works fast, without failure. So that's still a big win.

Also, I was able to delete a bunch of unused code.