Upgrade grid editor to work with Python 3.10. (#708)
Python 3.10 no longer auto-truncates floats to ints on the C backend.  This change explicitly changes all arguments of wx.Point to integers before passing in to solve this issue.

Relatedly, the outlines were not being shown, as the numpy arrays were no longer being properly cast to wx.Points. Thus, this explicitly builds the polygons as lists of wx.Points to solve that problem as well.

With these changes, the grid editor works in Python 3.10 as expected.