This adds a fix to the uniform mesh converter where data was being removed from the source reactor if it was not set on the converted reactor. This allows us to preserve the existing data on the blocks and reactor core if it is unchanged from any work that is performed on the converted reactor core.
Additionally, this moves the `axialMesh` parameter on the core outside
of a "neutronics" parameter category, as we want to avoid changing this
during the transfer of data from the uniform mesh reactor core back to
the non-uniform mesh reactor core. This is a significant bug that would
lead the reactor core state on non-uniform mesh to be inconsistent with
the core assembly meshes.