Number density bug fix (#804)
PR #657 introduced a bug into how number densities are initialized for a component.

The bug:

When `detailedAxialExpansion: True`, number densities were being initialized with input temperatures.
This is incorrect behavior as number densities should be initialized with hot temperatures to account for radial expansion of components. Otherwise, number densities are never adjusted for radial expansion and masses are incorrect.

This PR fixes this by initializing the number densities with hot temperatures. Accounting for axial expansion via inputHeightsConsideredHot is then accounted for in adjustNDensForHotHeight.

In addition to this fix, I've changed method names and improved docstrings in the interest clarity.
Also cleaned up the code within `adjustNDensForHotHeight` to be simpler and more efficient (instead of completely recalculating the number densities previously calculated in `setNDensFromMassFracsAtTempInC`, it now just scales them accordingly.