Add try/except around unpackSpecialData (#809)
compareDatabases should be able to continue to run even if unpackSpecialData fails.

Some parameters are not able to be compared, which in a recent scenario led to a failure inside unpackSpecialData. This PR adds the try/except so that there is just an error and not a failure, and it also adds a the traceback to the log printout so parameters that cause this step to fail can be more easily addressed / manually compared.