Parameter mapping fix in the uniform mesh converter. (#799)
* This adds a fix to the uniform mesh converter where data was being
removed from the source reactor if it was not set on the converted
reactor. This allows us to preserve the existing data on the blocks and
reactor core if it is unchanged from any work that is performed on
the converted reactor core.

Additionally, this moves the `axialMesh` parameter on the core outside
of a "neutronics" parameter category, as we want to avoid changing this
during the transfer of data from the uniform mesh reactor core back to
the non-uniform mesh reactor core. This is a significant bug that would
lead the reactor core state on non-uniform mesh to be inconsistent with
the core assembly meshes.

* Update the initialization of a new reactor to get name from the previous reactor passed in. This allows for two things: (1) further decoupling of the reactor and the operator for future refactoring and ease of running analyses without requiring an operator, and (2) not all reactors contain an instance of the operator so this can lead to failures.

* Fix getting case settings when operator is not available on the reactor.

* Improving the reporting of the cross section library that is attached
to the reactor core state.