Refactoring the `setStateFromOverlaps` to handle both the parameter mapping between assemblies as well as the number density mapping. This is now applied when calling `makeAssemWithUniformMesh`.
Update to the framework getIntegratedMgFlux to return 0.0 values rather
than a None if not set. Without this change, the refactor of the
uniform mesh was not functional. The change here is that if `calcReactionRates`
was provided previously, the source reactor would initialize the mgFlux parameters
to be an empty list rather than None. With this change, the source reactor parameters
that are unset remain as their defaults. The reaction rate calculation fails
if the mgFlux values on a composite is returned as `None` due to a `NoneType`
divide by float error.