Fix bug in rotatePins()
The assignment for pinLocation was broken. I am not sure why it was set
up the way it was, but it was using ParameterCollection.__setitem__() and
setting the `name` argument to the tuple ("pinLocation", pinNum). That
throws a TypeError from setattr(), which gets caught and then instead
makes an assigment to self._hist because `name` is a tuple. That is
definitely not what is intended.