Allow treatment of non-uniform assemblies in the uniform mesh converter (#879)
The uniform mesh module was updated to treat the conversion of non-uniform assemblies in the core model (using the nonUniformAssemFlags setting). In this change, a new branch of code is introduced within the UniformMeshConverter that short cuts the entire new reactor creation and entire reactor remeshing/data mapping when the nonUniformAssemFlags is provided by the user. This new feature allows for significant acceleration of the reactor unification because only a subset of all the assemblies need to be converted.