Fix the `getReactionRates()` methods to account for burnup, and also refactor
The primary purpose of this commit is to fix the `getReactionRates()` methods
on the `Block` and `Component` classes to account for burnup-dependent cross-sections.
This is completed primarily by changing the underlying `getReactionRateDict()`
method to accept the entire library suffix instead of just the first letter of
the suffix (i.e. "AA" or "AB" instead of just "A").

In the process, it was seen that a refactor could be achieved where the two classes
could share the method of the parent instead of implementing their own, since they
effectively do the same thing. This was accomplished by generalizing the calls
in the method and moving that method into the `Composite` class.