Use r.core.axialMesh() for uniform mesh.
When making a uniform mesh assembly in the UniformMeshConverter, the
freshly minted assembly does not have any submesh (i.e., b.p.axMesh ==
1). Homogenizing onto the refAssem's block mesh
 [r.core.refAssem.getAxialMesh()] poses problems if some blocks in
the refAssem have a submesh. r.core.axialMesh() contains the
computational neutronics mesh, which has all of the mesh subdivisions;
this is the mesh that should be used for makeAssemWithUniformMesh().