Fix numerical diffusion in uniformMesh converter
Mapping anything from a non-uniform assembly to a uniform assembly and
then back will cause numerical diffusion of that value. This could be a
block parameter (like mgFlux, detailedDpa, etc.) or number densities.

This commit deals with the issue of block parameters being mapped back
and forth. The number density issue has already been addressed for the
nonUniformAssemFlags case, but it seems to still be present for the
detailedAxialExpansion case.