Map power to uniform mesh.
The power isn't used for any real calculations, but it's necessary to
map power onto the uniform mesh reactor because some downstream physics
solvers want to check the energy balance before proceeding. The "energy
balance" is checked by comparing the sum of b.p.power to the power
specified by the settings input.