Updates to the fission product model and nuclides to allow for users to model fission products explicitly based on which cross section libraries/software will be used to generate multi-group cross sections.
This implements a method that looks at the existing nuclides defined
in the blueprints `nuclideFlags` and amends this to include all nuclides
within regions that are determined to be `Depletable` based on which nuclides
the user sets as being `burn=True`. This will not check that the user
defines the correct depletable nuclides for this determination, but will
only add all the nuclides that are missing to the depletable regions. This
is reasonable since the behavior is similar to how it is when the user
selects another fission product model option.

This change also introduces a mechanism to remove gaseous fission products
from the core that can be called upon by a fuel performance or other
plugin code. The previous implementation of this modified the fission
yields as a surrogate for this removal, but upon further review it is
better to update the number densities within the blocks directly compared
to modifying the fission yields. By modifying the number densities directly
we are explicitly handling both the microscope and macroscopic cross section
effects in lattice physics calculations and in system eigenvalue/flux
solutions.

Finally, the fuel performance parameters were updated to add setters on
the `gasReleaseFraction` and the `bondRemoved`. The description suggested
that these values should only range from [0, 1] and any value outside of
this range is not valid. The setters now check for this and raise a
ValueError if these conditions are not met.