Extension of the framework nuclides (#998)
Update to the nuclides that are built into the ARMI framework. This expands the set of nuclide bases to include 4614 nuclides and imports the half-life data from the chart of the nuclides from a combination of the RIPL-3 database and from the IAEA chart of the nuclides.

Other Changes:

Each Element now has a phase and a group.
Element.nuclideBases becomes Element.nuclides - This is an API-breaking change.