Rename block.createCopy() and set number densities
Rename block.createCopy() to block._createHomogenizedCopy() to
denote what it is doing in more detail, and make it "private" to
indicate that it generally shouldn't be used in scripts as part
of the API. Even though it was made "private", It is still called
in the uniform mesh converter.

Setting number densities to equal the source block number densities
to make _createHomogenized copy return a block with number densities in
all cases.