Add flag on converter to include pin coordinates.
It is significantly more expensive to include the original
pin component spatialLocators (i.e., pin coordinates) when
creating a homogenized copy of a block. A lot of computational
effort can be saved by not copying this data over unless it
is going to be used by the physics solver.

Also, support multiple pin types per block.