Add armi.reactor.compositeList.CompositeList : base class for Assemblies
The intention of the abstract parent class is to shift most of the composite-tree
traversal routines onto a generic parent class. It's abstract only because the
Assembly does some re-naming of children during `Assembly.add` which subclasses
might want to override. Might not be enough to justify all the metaclasses and just
make that an empty method.

Later, another subclass of `CompositeList` could be used to model stacks of fuel
pellets in a ``Block``. But that would require many blueprint side modifications
that are to be discussed.

Some assembly methods like `getBlockAndZ` pass directly off to the parent method
to maintain the same API. Wondering if there is room to make some of these
iterator-based methods a generator? But that would also be an API change and should
be done after some discussions.

Added a test for getBlocksAtElevation.

Closes https://github.com/terrapower/armi/issues/1030

Signed-off-by: Andrew Johnson <a.johnson@usnc-tech.com>