Efficiency improvement for uniform mesh converter and general operations (#1042)
Efficiency improvements for uniform mesh converter
--------------------------------------------------
Various operations that are involved in the uniform mesh 
conversion process were updated for better efficiency. The 
targets for optimization were identified through profiling a 
sample uniform mesh conversion. Several of the functions that 
were sped up through this process will also improve speed in 
general across a broad range of ARMI operations, including 
initialization of the model from blueprints, performing axial 
expansion, loading from a database, etc.

Speed up the following functions:

* composite.getNuclideNumberDensities()
* axialExpChngr._conserveComponentDensity()
* UniformMeshGeometryConverter.setAssemblyStateFromOverlaps()
* globalFlux.calcReactionRates()
* GlobalFluxInterfaceUsingExecuters.calculateKeff() (skip calcReactionRates)

Additional Notes on Updates
---------------------------
* Use Block._createHomogenizedCopy() method to copy the basic features of a block
without using deepcopy. This is only implemented for a HexBlock at this point.
Still needs to use a deepcopy of the components for pin-level spatial information, 
but we could replace that with a custom blank copy function for components. 
It is named like a private function to indicate that it generally shouldn't be used 
in scripts as part of the API. Even though it was made "private", It is still called
in the uniform mesh converter.

* _calculateReactionRates() operates on the whole destination assembly,
but it was incorrectly positioned inside of the loop over blocks within
the assembly, creating an unnecessary double-nested loop over blocks
within an assembly. This was a significant factor slowing down the
uniform mesh converter.