Cleanup umc (#1099)
power and pdens parameters are initially set in a neutronics run,
but they may be adjusted by a gamma transport solver. Thus, these
parameters should be mapped out of a uniform mesh gamma solve, even
though they are also mapped in. If a gamma kernel is implemented that
does not update these parameters, this converter would have to be
modified and/or extended to prevent numerical diffusion of these
parameters.

Also, set b.p.nPins in Block._createHomogenizedCopy().