Adding measure of convergence for tight coupling  (#1033)
This PR adds the capability to quantifiably measure the numerical convergence of solving nonlinear systems of equations when considering tight coupling. The current approach within ARMI just iterates on the system of equations using the numCoupledIterations setting. While indispensable for framework development, characterizing the numerical convergence of the nonlinear system of equations is essential to ensuring the mathematical validity of obtained solutions.

Co-authored-by: jakehader <jhader@terrapower.com>