Refactoring Material to not subclass Composite (#1062)
This PR refactors the Material class to no longer be a subclass of Composite.

It also removes (most, not all) of the statefulness in ARMI Materials. For instance, I removed all Parameters from Material. Some of those I copied over to a simpler class attribute (mat.p.massFrac became mat.massFrac). And some I was able to delete entirely (like yieldStrength and atomFracDenom).

Also, I should note that three parameters (uFrac, puFrac, and zrFrac) applied only to FUEL materials, but were defined by default for all materials. That was fixed.

Stateful things I haven't removed in this PR (because they are going to be tricky):

* ARMI supports Custom materials, which are obviously stateful by design.
* ARMI supports custom isotopics for fuels, which again is a stateful idea.
* The Material class still has a self.parent, which is necessary for a whole LOAD of logic to do with Material.getNuclides().