Compute uniform axial mesh with k-means
When running the uniform mesh converter, we need a way to determine the
best mesh onto which the non-uniform assemblies will be homogenized. The
method of _computeAverageAxialMesh just averages all of the meshes of
the different assemblies and tosses out any assembly that isn't within
tolerance of the average mesh. This doesn't do a good job of preserving
mesh lines at important material boundaries.

A better method for preserving more mesh lines is to use k-means
clustering to determine an appropriate mesh that matches the most
important mesh lines. The new method uses k-means with an increasing
number of mesh points until it begins making meshes that are finer than
the mesh tolerance (default = 1.0 cm).