Calculate block masses at BOC
Block heavy metal and fissile masses are calculated in the depletion
interface, which is the only interface where they would typically be
affected. However, depletion is typically inactive at node 0 of a cycle.
This PR adds a simple calculation for b.p.kgHM and b.p.kgFis on core
loading and after shuffling.