Add Copying Existing Flux Solutions to Cross Section Control Settings (#1152)
* Adding new functionality to the cross section settings to copy
existing flux solution files during lattice physics evaluations.
This is useful for lattice physics codes that allow for an external
solver solution file to be provided.

* This breaks the crossSectionControl setting for existing applications
with the use of pregenerated cross sections. The name of the setting
was changed from `fileLocation` to `xsFileLocation` to allow the addition
of the `fluxFileLocation` setting.