add a check to prevent multiple loads of userPlugins
During the submission of cases, userPlugins are registered. This can result
in multiple attempts to load a specific userPlugin if a case-suite is
being submitted, as each loading of the input yaml attempts to register
the userPlugin.

A check has been added to skip the registration of the userPlugin if a
userPlugin with the same name already exists in the registry.