Fixing UserPlugins during CaseSuite (#1183)
During the submission of cases, UserPlugins are registered. This can result
in multiple attempts to load a specific UserPlugin if a case-suite is
being submitted, as each loading of the input yaml attempts to register
the UserPlugin.

A check has been added to skip the registration of the UserPlugin if a
UserPlugin with the same name already exists in the registry.