Control decusping (#1229)
The performance of the UniformMeshGeometryConverter depends significantly on how good of a common uniform mesh can be generated. The uniform mesh converter will homogenize materials in neighboring blocks, which can be a problem at material boundaries between fuel or absorber and another material. One manifestation of this issue is a significant control rod "cusping" effect that has been observed when calculating differential rod worth.

The existing method _computeAverageAxialMesh() is too simplistic and is totally agnostic to material boundaries. This PR introduces a new class, UniformMeshGenerator, that has a set of methods to modify the mesh produced by _computeAverageAxialMesh so that it does a better job of representing important material boundaries. The main goal of this PR is to eliminate control rod cusping effects, but it will also help when modeling a core that has fuel assemblies at different heights.