Add 1D Cylindrical Modeling to the XS Group Manager (#1238)
Implement a block collection type for managing 1D cylindrical
control rod modeling to ensure that number densities do not migrate
to other components. This is avoided by building a representative
block and setting number densities directly onto the components
instead of using `b.setNumberDensities()`. Setting number densities
directly on the blocks is not a good technique if the blocks are to
be treated heterogeneously because there is a chance that material
will migrate to other regions based on the "dehomogenization" logic
that is applied.

Co-authored-by: Michael Jarrett <mjarrett@terrapower.com>