Composite.getBoundingCircleOuterDiameter sums over children, not iterComponents
For particle fuel, we potentially could have very small spheres that show
up in `iterComponents` that would be contained inside a larger matrix component.
Here, we want the "bounding diameter" to only be that of the matrix as, physically,
the diameter of the matrix bounds the compact.

This allows intermediate composites like the `Compact` to override their own
implementation that just returns the bounding diameter of the matrix