Refactor armi.reactor.grids to wider submodule; Grids as ABC (#1373)
* Move grids.py to own module armi.reactor.grids

Goal is to introduce separate files for each grid type, but still
make those importable from armi.reactor.grids

* Add armi.reactor.grids.constants

Intended to share some constants and type hints across the
`armi.reactor.grids` submodule. Intentionally sparse.

Items will be imported into `armi.reactor.grids` to help with
back compatibility.

* Add armi.reactor.grids.locations

Provides abstract `LocationBase` as the interface defining base class.
Concrete subclasses `IndexLocation`, `MultiIndexLocation`, and
`CoordinateLocation` are also defined.

There is little to no implementation differences in this change. Only
that `LocationBase` is a proper abstract base class, so subclasses that
don't override the necessary abstract methods will need to change. Some
other key changes

- `LocationBase.__eq__` returns `NotImplemented` if the RHS argument
  is not a tuple nor a `LocationBase`

The four main classes defined are imported into `armi.reactor.grids`
to maintain back compatibility.

* Move reactor/tests/test_grids to grids/tests

* Add armi.reactor.grids.grid for Grid base class

Imported into `armi.reactor.grids` so still back compatible

* Add armi.reactor.grids.cartesian for CartesianGrid

* Add armi.reactor.grids.hexagonal

* Add armi.reactor.grids.thetarz

* Remove six.cPickle -> pickle in test_grids.py

* Add armi.reactor.grids.axial with AxialGrid class

Replaces grids.axialUnitGrid

* Move duplicated grids.addingIsValid to grids.locations

* Refactor Grid to be ABC, intermediate StructuredGrid

Heavier lift, but the classes defined in `armi.reactor.grids`
follow this inheritance pattern

- `Grid(abc.ABC)` :: abstract interface for _a_ grid, or things that reside in
  space
  - `StructuredGrid(Grid)` :: still abstract, but for things that have
    some structure, known to ARMI as unit steps and bounds
    - `HexGrid(StructuredGrid)`
    - `CartesianGrid(StructuredGrid)
    - `ThetaRZGrid(StructuredGrid)` - though how "structured" is this
      really?
    - `AxialGrid(StructuredGrid)`

In some cases, the shared interface between structured grids falls
apart. For example, `HexGrid.pitch` returns a single float, while
`CartesianGrid.pitch` is a 2-ple of float, one for x and one for y.
And the concept of ring and position appears to be stressed since the
inception of cartesian grids. But parts of ARMI require an interface
and so an attempt was made to define that interface.

The decision was made to not have `StructuredGrid` the base class, and
to introduce `Grid` as the base intentionally. There are cases for some
multiphysics codes where a grid may still be "structured" but not with
the level of structure the previous `Grid` implementations expect. One
example would be an irregular cartesian grid, where the unit cell
spacing in one or more dimensions may change. So the cell spacing is a
function of the cell position in that dimension.