Activate sphinx-needs extension in doc config.
The docs have lots of sphinx needs directives in them, including
implementation (.. impl::) and tests (.. test::). These are defined by
the sphinx-needs extension, which is in our requirements but not
actually activated in our docs config. This activates it in the docs.
The net result is that hundreds of errors in the doc build process are
resolved.