Adding more prominent grid editor docs (#1731)
The grid editor is a powerful feature, and not easy
to discover or featured in the docs. This adds
an intriguing figure (from our fftf isothermal repo)
up front so people can see what they're missing
and go looking for the feature.

Along the way, I noticed that all png files were
ignored across the repo. Since we use png images
in our documentation, it's important that we do not
blanket ignore such files. I struggled for a while
to find out why my png wasn't showing up in the options
to stage! If we have build artifacts that are ending
up in commits then I suggest we clean those out
or be more precise with the ignoring.