remove test for bad custom materials in tutorial input
the tutorial input was assigning custom isotopics to a component
made of void material. This was being used to check that warnings
were being logged that the material properties were ignored.
Allowing isotopics to be assigned to Void material is not allowed
anymore and an error (not a warning) is thrown. The testing for
material properties being added to void material have been
implemented elsewhere.